Create UpdateApkClient
